# ENH: improve performance of `SphericalVoronoi` area calculation