* the 2D (circle-based) `SphericalVoronoi` arc area calculation seemed a bit complex to me, so I tried simplifying it, and it seems to perform better when benchmarks are expanded to include 2D "area" calcs

* `asv continuous -e -b "time_spherical_polygon_area_calculation" main treddy_area_2d_simplify`

```
       before           after         ratio
     [bcf81dd3]       [796c0660]
     <main>           <treddy_area_2d_simplify>
-      17.9±0.1μs      13.2±0.04μs     0.73  spatial.SphericalVorAreas.time_spherical_polygon_area_calculation(100, 2)
-        97.1±2μs       48.8±0.6μs     0.50  spatial.SphericalVorAreas.time_spherical_polygon_area_calculation(1000, 2)
-       428±0.6μs          201±5μs     0.47  spatial.SphericalVorAreas.time_spherical_polygon_area_calculation(5000, 2)
-      1.01±0.1ms          390±8μs     0.38  spatial.SphericalVorAreas.time_spherical_polygon_area_calculation(10000, 2)

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```